Move methods where they belong to
