+++ src/Logging/JUnit/Reader.php
@@ -231,4 +231,42 @@ class Reader extends MetaProvider
             $this->suites[] = TestSuite::suiteFromArray(self::$defaultSuite);
         }
     }
+
+    /**
+     * Return a value as a float or integer.
+     *
+     * @param $property
+     *
+     * @return float|int
+     */
+    protected function getNumericValue(string $property)
+    {
+        return ($property === 'time')
+            ? (float) $this->suites[0]->$property
+            : (int) $this->suites[0]->$property;
+    }
+
+    /**
+     * Return messages for a given type.
+     *
+     * @param $type
+     *
+     * @return array
+     */
+    protected function getMessages(string $type): array
+    {
+        $messages = [];
+        $suites = $this->isSingle ? $this->suites : $this->suites[0]->suites;
+        foreach ($suites as $suite) {
+            $messages = \array_merge($messages, \array_reduce($suite->cases, function ($result, $case) use ($type) {
+                return \array_merge($result, \array_reduce($case->$type, function ($msgs, $msg) {
+                    $msgs[] = $msg['text'];
+
+                    return $msgs;
+                }, []));
+            }, []));
+        }
+
+        return $messages;
+    }
 }
